is_valid_date

+++ cpp09/ex00/includes/BitcoinExchange.hpp
@@ -6,7 +6,7 @@
 /*   By: mfeldman <mfeldman@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/15 17:18:54 by mfeldman          #+#    #+#             */
-/*   Updated: 2024/04/21 18:58:53 by mfeldman         ###   ########.fr       */
+/*   Updated: 2024/04/22 22:10:50 by mfeldman         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,6 +30,6 @@ class BitcoinExchange
 		
 	private:
     	std::string _filename;
-		std::map<std::string, float> _bitcoin_data;
+		std::map<std::string, double> _bitcoin_data;
 };
